requirements.txt
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.6
fastapi==0.100.0
frozenlist==1.4.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
multidict==6.0.4
Pillow==10.0.0
pydantic==2.0.3
pydantic_core==2.3.0
python-dotenv==1.0.0
python-telegram-bot==20.4
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
urllib3==2.0.3
uvicorn==0.23.1
yarl==1.9.2
aiohttp==3.8.5